+++ app/datatables/skaters_datatable.rb
@@ -3,10 +3,13 @@ class SkatersDatatable < IndexDatatable
     super
 
     columns([:name, :category, :nation, :isu_number, :birthday, :club, :coach])
-    default_orders([[:category, :asc], [:name, :asc]])
+    #default_orders([[:category, :asc], [:name, :asc]])
+    columns.sources = {
+      category: "categories.name",
+    }
   end
 
   def fetch_records
-    Skater.having_scores
+    Skater.having_scores.includes(:category).references(:category)
   end
 end