+++ spec/factories/factories.rb
@@ -1,20 +1,24 @@
 # skaters
 FactoryBot.define do
   factory :skater do
+    association :category, factory: :category
+
     name { "Taro YAMADA" }
     nation { "JPN" }
     isu_number { 1 }
-    category { "MEN" }
+    #category { "MEN" }
+    
     birthday { "1.1.1980" } 
     coach { "Jiro YAMADA" }
     hometown { "TOKYO" }
     club { "Jingu" } 
 
     trait :ladies do
+      association :category, factory: [:category, :ladies]
       name { "Leia ORGANA" }
       nation { "USA" }
       isu_number { 2 }
-      category { "LADIES" }
+      #category { "LADIES" }
       birthday { "1.1.1990" }
       coach { "Rola ORGANA" }
       hometown { "L.A." }
@@ -28,6 +32,7 @@ FactoryBot.define do
     name { "MEN" }
     abbr { "SM" }
     seniority { "SENIOR" }
+    indivisual { true }
 
     trait :ladies do
       name { "LADIES" }